Move out scrapeSingleArticleInfo
