+++ src/scrape/scrapeAllReadings.ts
@@ -1,10 +1,9 @@
 import cliProgress from 'cli-progress';
-import { JSDOM } from 'jsdom';
 import { PIXIV_BASE_URL } from '../constants';
-import { fetchURL } from '../fetch/fetchURL';
 import { prisma } from '..';
+import { scrapeSingleArticleInfo } from './scrapeSingleArticleInfo';
 
-const pixivArticleURL = (tag_name: string) =>
+export const pixivArticleURL = (tag_name: string) =>
   `${PIXIV_BASE_URL}a/${encodeURIComponent(tag_name)}`;
 /**
  * Scrape all readings for articles that have not been scraped yet
@@ -37,28 +36,4 @@ export async function scrapeAllReadings() {
     progressBar.update(i);
   }
   progressBar.stop();
-}
-
-async function scrapeSingleArticleInfo(tag_name: string) {
-  // Fetch the page
-  const url = pixivArticleURL(tag_name);
-  const response = await fetchURL(url);
-  const dom = new JSDOM(response.data);
-  // Get the reading
-  const reading = dom.window.document.querySelector('p.subscript')?.textContent;
-  // Get headers
-  const breadcrumbs = dom.window.document.getElementById('breadcrumbs');
-  const header = breadcrumbs
-    ? [...breadcrumbs.children].map((child) => child.textContent)
-    : [];
-
-  // Update the article
-  await prisma.pixivArticle.update({
-    where: { tag_name },
-    data: {
-      lastScrapedReading: Date.now().toString(),
-      reading,
-      header: JSON.stringify(header),
-    },
-  });
 }
